Extract CSS vars from RAC docs into reusable package
