+++ packages/@react-aria/example-theme/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@react-aria/example-theme",
+  "version": "1.0.0",
+  "description": "Spectrum UI components in React",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/adobe/react-spectrum"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
