+++ packages/react-aria-components/docs/Button.mdx
@@ -50,18 +50,11 @@ import {Button} from 'react-aria-components';
   <summary style={{fontWeight: 'bold'}}><ChevronRight size="S" /> Show CSS</summary>
 
 ```css
-.react-aria-Button {
-  --border-color: var(--spectrum-alias-border-color);
-  --border-color-pressed: var(--spectrum-alias-border-color-down);
-  --border-color-disabled: var(--spectrum-alias-border-color-disabled);
-  --background-color: var(--spectrum-global-color-gray-50);
-  --background-color-pressed: var(--spectrum-global-color-gray-100);
-  --text-color: var(--spectrum-alias-text-color);
-  --text-color-disabled: var(--spectrum-alias-text-color-disabled);
-  --focus-ring-color: slateblue;
+@import "@react-aria/example-theme/index.css";
 
+.react-aria-Button {
   color: var(--text-color);
-  background: var(--background-color);
+  background: var(--button-background);
   border: 1px solid var(--border-color);
   border-radius: 4px;
   appearance: none;
@@ -75,27 +68,13 @@ import {Button} from 'react-aria-components';
 
   &[data-pressed] {
     box-shadow: inset 0 1px 2px rgb(0 0 0 / 0.1);
-    background: var(--background-color-pressed);
+    background: var(--button-background-pressed);
     border-color: var(--border-color-pressed);
   }
 
   &[data-focus-visible] {
-    border-color: var(--focus-ring-color);
-    box-shadow: 0 0 0 1px var(--focus-ring-color);
-  }
-}
-
-@media (forced-colors: active) {
-  .react-aria-Button {
-    forced-color-adjust: none;
-    --border-color: ButtonBorder;
-    --border-color-pressed: ButtonBorder;
-    --border-color-disabled: GrayText;
-    --background-color: ButtonFace;
-    --background-color-pressed: ButtonFace;
-    --text-color: ButtonText;
-    --text-color-disabled: GrayText;
-    --focus-ring-color: Highlight;
+    outline: 2px solid var(--focus-ring-color);
+    outline-offset: -1px;
   }
 }
 ```
