+++ packages/react-aria-components/docs/Calendar.mdx
@@ -62,16 +62,9 @@ import {Calendar, Heading, Button, CalendarGrid, CalendarCell} from 'react-aria-
 ```
 
 ```css
-.react-aria-Calendar {
-  --highlight-background: slateblue;
-  --highlight-foreground: white;
-  --text-color: var(--spectrum-alias-text-color);
-  --text-color-disabled: var(--spectrum-alias-text-color-disabled);
-  --text-color-invalid: var(--spectrum-global-color-red-600);
-  --pressed-color: var(--spectrum-global-color-gray-200);
-  --unavailable-color: var(--spectrum-global-color-red-600);
-  --invalid-color: var(--spectrum-global-color-static-red-600);
+@import "@react-aria/example-theme/index.css";
 
+.react-aria-Calendar {
   width: fit-content;
   max-width: 100%;
   color: var(--text-color);
@@ -102,41 +95,26 @@ import {Calendar, Heading, Button, CalendarGrid, CalendarCell} from 'react-aria-
     border-radius: 6px;
     cursor: default;
     outline: none;
-    border: 2px solid var(--spectrum-alias-background-color-default);
-    margin: -1px;
+    margin: 1px;
+    forced-color-adjust: none;
 
     &[data-outside-month] {
       display: none;
     }
 
     &[data-pressed] {
-      background: var(--pressed-color);
+      background: var(--gray-100);
     }
 
     &[data-focus-visible] {
-      box-shadow: 0 0 0 2px var(--highlight-background);
+      outline: 2px solid var(--focus-ring-color);
+      outline-offset: 2px;
     }
 
     &[data-selected] {
       background: var(--highlight-background);
       color: var(--highlight-foreground);
     }
-  }
-}
-
-@media (forced-colors: active) {
-  .react-aria-Calendar {
-    forced-color-adjust: none;
-
-    --highlight-background: Highlight;
-    --highlight-foreground: HighlightText;
-    --text-color: ButtonText;
-    --text-color-disabled: GrayText;
-    --text-color-invalid: LinkText;
-    --unavailable-color: GrayText;
-    --invalid-color: LinkText;
-    --pressed-color: Canvas;
-    --spectrum-alias-background-color-default: Canvas;
   }
 }
 ```
@@ -430,7 +408,7 @@ function Example() {
   .react-aria-CalendarCell {
     &[data-unavailable] {
       text-decoration: line-through;
-      color: var(--unavailable-color);
+      color: var(--invalid-color);
     }
   }
 }
@@ -478,9 +456,10 @@ function Example() {
       color: var(--highlight-foreground);
     }
   }
+
   [slot=errorMessage] {
     font-size: 12px;
-    color: var(--text-color-invalid);
+    color: var(--invalid-color);
   }
 }
 ```
