+++ packages/react-aria-components/docs/Checkbox.mdx
@@ -57,25 +57,23 @@ import {Checkbox} from 'react-aria-components';
   <summary style={{fontWeight: 'bold'}}><ChevronRight size="S" /> Show CSS</summary>
 
 ```css
+@import "@react-aria/example-theme/index.css";
+
 .react-aria-Checkbox {
-  --label-color: var(--spectrum-alias-text-color);
-  --deselected-color: gray;
-  --deselected-color-pressed: dimgray;
-  --selected-color: slateblue;
-  --selected-color-pressed: lch(from slateblue calc(l - 10%) c h);
-  --checkmark-color: white;
-  --invalid-color: var(--spectrum-global-color-red-600);
-  --invalid-color-pressed: var(--spectrum-global-color-red-700);
+  --selected-color: var(--highlight-background);
+  --selected-color-pressed: var(--highlight-background-pressed);
+  --checkmark-color: var(--highlight-foreground);
 
   display: flex;
   align-items: center;
   gap: 0.571rem;
   font-size: 1.143rem;
+  forced-color-adjust: none;
 
   .checkbox {
     width: 1.143rem;
     height: 1.143rem;
-    border: 2px solid var(--deselected-color);
+    border: 2px solid var(--border-color);
     border-radius: 4px;
     transition: all 200ms;
     display: flex;
@@ -83,7 +81,7 @@ import {Checkbox} from 'react-aria-components';
     justify-content: center;
   }
 
-  & svg {
+  svg {
     width: 1rem;
     height: 1rem;
     fill: none;
@@ -95,12 +93,14 @@ import {Checkbox} from 'react-aria-components';
   }
 
   &[data-pressed] .checkbox {
-    border-color: var(--deselected-color-pressed);
+    border-color: var(--border-color-pressed);
   }
 
   &[data-focus-visible] .checkbox {
-    box-shadow: 0 0 0 2px var(--spectrum-alias-background-color-default), 0 0 0 4px var(--selected-color);
+    outline: 2px solid var(--focus-ring-color);
+    outline-offset: 2px;
   }
+
   &[data-selected],
   &[data-indeterminate] {
     .checkbox {
@@ -113,7 +113,7 @@ import {Checkbox} from 'react-aria-components';
       background: var(--selected-color-pressed);
     }
 
-    & svg {
+    svg {
       stroke-dashoffset: 44;
     }
   }
@@ -122,28 +122,6 @@ import {Checkbox} from 'react-aria-components';
     & svg {
       stroke: none;
       fill: var(--checkmark-color);
-    }
-  }
-}
-
-@media (forced-colors: active) {
-  .react-aria-Checkbox {
-    forced-color-adjust: none;
-
-    --label-color: ButtonText;
-    --deselected-color: ButtonBorder;
-    --deselected-color-pressed: ButtonBorder;
-    --selected-color: Highlight;
-    --selected-color-pressed: Highlight;
-    --checkmark-color: HighlightText;
-    --invalid-color: LinkText;
-    --invalid-color-pressed: LinkText;
-    --spectrum-alias-background-color-default: Canvas;
-
-    &[data-disabled] {
-      --deselected-color: GrayText;
-      --selected-color: GrayText;
-      --label-color: GrayText;
     }
   }
 }
@@ -264,7 +242,7 @@ Implement your own validation logic in your app and set the `isInvalid` prop acc
 .react-aria-Checkbox {
   &[data-invalid] {
     .checkbox {
-      --checkmark-color: var(--spectrum-gray-50);
+      --checkmark-color: var(--gray-50);
       border-color: var(--invalid-color);
     }
 
@@ -301,14 +279,10 @@ Checkboxes can be disabled using the `isDisabled` prop.
 ```css
 .react-aria-Checkbox {
   &[data-disabled] {
-    opacity: 0.4;
-  }
-}
+    color: var(--text-color-disabled);
 
-@media (forced-colors: active) {
-  .react-aria-Checkbox {
-    &[data-disabled] {
-      opacity: 1;
+    .checkbox {
+      border-color: var(--border-color-disabled);
     }
   }
 }
