+++ packages/react-aria-components/docs/CheckboxGroup.mdx
@@ -72,6 +72,7 @@ import {CheckboxGroup, Checkbox, Label} from 'react-aria-components';
 <details>
   <summary style={{fontWeight: 'bold'}}><ChevronRight size="S" /> Show CSS</summary>
 ```css hidden
+@import "@react-aria/example-theme/index.css";
 @import './Checkbox.mdx' layer(checkbox);
 @import './Form.mdx' layer(form);
 @import './Button.mdx' layer(button);
@@ -295,13 +296,7 @@ import {Form, FieldError, Button} from 'react-aria-components';
 ```css
 .react-aria-FieldError {
   font-size: 12px;
-  color: var(--spectrum-global-color-red-600);
-}
-
-@media (forced-colors: active) {
-  .react-aria-FieldError {
-    color: LinkText;
-  }
+  color: var(--invalid-color);
 }
 ```
