+++ packages/react-aria-components/docs/Dialog.mdx
@@ -73,6 +73,7 @@ import {DialogTrigger, Modal, Dialog, Button, Heading, TextField, Label, Input} 
 <details>
   <summary style={{fontWeight: 'bold'}}><ChevronRight size="S" /> Show CSS</summary>
 ```css hidden
+@import "@react-aria/example-theme/index.css";
 @import './Button.mdx' layer(button);
 @import './TextField.mdx' layer(textfield);
 @import './Popover.mdx' layer(popover);
