+++ packages/react-aria-components/docs/FileTrigger.mdx
@@ -62,6 +62,7 @@ function Example(){
 ```
 
 ```css hidden
+@import "@react-aria/example-theme/index.css";
 @import './Button.mdx' layer(button);
 ```
