+++ packages/react-aria-components/docs/Link.mdx
@@ -52,13 +52,10 @@ import {Link} from 'react-aria-components';
   <summary style={{fontWeight: 'bold'}}><ChevronRight size="S" /> Show CSS</summary>
 
 ```css
-.react-aria-Link {
-  --focus-ring-color: slateblue;
-  --text-color: var(--spectrum-global-color-blue-600);
-  --text-color-pressed: var(--spectrum-global-color-blue-700);
-  --text-color-disabled: gray;
+@import "@react-aria/example-theme/index.css";
 
-  color: var(--text-color);
+.react-aria-Link {
+  color: var(--link-color);
   font-size: 18px;
   transition: all 200ms;
   text-decoration: underline;
@@ -71,7 +68,7 @@ import {Link} from 'react-aria-components';
   }
 
   &[data-pressed] {
-    color: var(--text-color-pressed);
+    color: var(--link-color-pressed);
   }
 
   &[data-focus-visible]:after {
@@ -80,15 +77,6 @@ import {Link} from 'react-aria-components';
     inset: -3px -6px;
     border-radius: 6px;
     border: 2px solid var(--focus-ring-color);
-  }
-}
-
-@media (forced-colors: active) {
-  .react-aria-Link {
-    --focus-ring-color: Highlight;
-    --text-color: LinkText;
-    --text-color-pressed: LinkText;
-    --text-color-disabled: GrayText;
   }
 }
 ```
