+++ packages/react-aria-components/docs/ListBox.mdx
@@ -64,14 +64,9 @@ import {ListBox, ListBoxItem} from 'react-aria-components';
 ```
 
 ```css
-.react-aria-ListBox {
-  --highlight-background: slateblue;
-  --highlight-foreground: white;
-  --border-color: var(--spectrum-global-color-gray-400);
-  --background-color: var(--page-background);
-  --text-color: var(--spectrum-alias-text-color);
-  --text-color-disabled: var(--spectrum-alias-text-color-disabled);
+@import "@react-aria/example-theme/index.css";
 
+.react-aria-ListBox {
   display: flex;
   flex-direction: column;
   max-height: inherit;
@@ -79,7 +74,8 @@ import {ListBox, ListBoxItem} from 'react-aria-components';
   padding: 2px;
   border: 1px solid var(--border-color);
   border-radius: 6px;
-  background: var(--background-color);
+  background: var(--overlay-background);
+  forced-color-adjust: none;
   outline: none;
   width: 250px;
   max-height: 300px;
@@ -87,8 +83,8 @@ import {ListBox, ListBoxItem} from 'react-aria-components';
   box-sizing: border-box;
 
   &[data-focus-visible] {
-    border-color: var(--highlight-background);
-    box-shadow: 0 0 0 1px var(--highlight-background);
+    outline: 2px solid var(--focus-ring-color);
+    outline-offset: -1px;
   }
 }
 
@@ -105,7 +101,8 @@ import {ListBox, ListBoxItem} from 'react-aria-components';
   flex-direction: column;
 
   &[data-focus-visible] {
-    box-shadow: inset 0 0 0 2px var(--highlight-background);
+    outline: 2px solid var(--focus-ring-color);
+    outline-offset: -2px;
   }
 
   &[data-selected] {
@@ -113,22 +110,9 @@ import {ListBox, ListBoxItem} from 'react-aria-components';
     color: var(--highlight-foreground);
 
     &[data-focus-visible] {
-      box-shadow: inset 0 0 0 2px var(--highlight-background), inset 0 0 0 4px var(--highlight-foreground);
+      outline-color: var(--highlight-foreground);
+      outline-offset: -4px;
     }
-  }
-}
-
-@media (forced-colors: active) {
-  .react-aria-ListBox {
-    forced-color-adjust: none;
-
-    --highlight-background: Highlight;
-    --highlight-foreground: HighlightText;
-    --border-color: ButtonBorder;
-    --background-color: ButtonFace;
-    --separator-color: ButtonBorder;
-    --text-color: ButtonText;
-    --text-color-disabled: GrayText;
   }
 }
 ```
@@ -244,9 +228,9 @@ function MyItem(props: ListBoxItemProps) {
   max-height: inherit;
   overflow: auto;
   padding: 2px;
-  border: 1px solid var(--spectrum-global-color-gray-400);
+  border: 1px solid var(--border-color);
   border-radius: 6px;
-  background: var(--page-background);
+  background: var(--overlay-background);
   outline: none;
   max-width: 250px;
   max-height: 300px;
@@ -261,33 +245,24 @@ function MyItem(props: ListBoxItemProps) {
   border-radius: 6px;
   outline: none;
   cursor: default;
-  color: var(--spectrum-global-color-gray-800);
+  color: var(--text-color);
   font-size: 1.072rem;
 
   &.selected {
     background: var(--highlight);
-    color: white;
+    color: var(--highlight-foreground);
   }
 
   &.focused {
-    box-shadow: 0 0 0 2px var(--page-background), 0 0 0 4px var(--highlight);
+    outline: 2px solid var(--highlight);
+    outline-offset: 2px;
   }
 }
 
 @media (forced-colors: active) {
-  .my-listbox {
-    forced-color-adjust: none;
-    border-color: ButtonBorder;
-    background: Canvas;
-  }
-
   .my-item {
-    color: ButtonText;
+    forced-color-adjust: none;
     --highlight: Highlight;
-
-    &.selected {
-      color: HighlightText;
-    }
   }
 }
 ```
@@ -629,10 +604,6 @@ let albums = [
   <summary style={{fontWeight: 'bold'}}><ChevronRight size="S" /> Show CSS</summary>
 
 ```css
-.react-aria-ListBox {
-  --selectedFilled: slateblue;
-  --selectedOutline: white;
-}
 .react-aria-ListBox[data-orientation=horizontal],
 .react-aria-ListBox[data-layout=grid] {
   flex-direction: row;
@@ -675,9 +646,9 @@ let albums = [
         position: absolute;
         top: 8px;
         right: 8px;
-        background: var(--selectedFilled);
-        border: 2px solid var(--selectedOutline);
-        color: var(--selectedOutline);
+        background: var(--highlight-background);
+        border: 2px solid var(--highlight-foreground);
+        color: var(--highlight-foreground);
         width: 22px;
         height: 22px;
         border-radius: 22px;
@@ -689,18 +660,7 @@ let albums = [
         justify-content: center;
         box-shadow: 0 0 8px rgb(0 0 0 / .5);
       }
-
-      &[data-focus-visible] {
-        box-shadow: inset 0 0 0 2px var(--highlight-background);
-      }
     }
-  }
-}
-
-@media (forced-colors: active) {
-  .react-aria-ListBox {
-    --selectedFilled: Highlight;
-    --selectedOutline: ButtonText;
   }
 }
 ```
@@ -1054,13 +1014,13 @@ function Example() {
   align-items: center;
   justify-content: space-between;
   gap: 4px;
-  background: slateblue;
-  color: white;
+  background: var(--highlight-background);
+  color: var(--highlight-foreground);
   border-radius: 4px;
 
   .badge {
-    background: white;
-    color: slateblue;
+    background: var(--highlight-foreground);
+    color: var(--highlight-background);
     padding: 0 8px;
     border-radius: 4px;
   }
@@ -1167,9 +1127,9 @@ function Example() {
 
 ```css
 .react-aria-ListBox[data-drop-target] {
-  border-color: var(--highlight-background);
-  box-shadow: 0 0 0 1px var(--highlight-background);
-  background: rgb(from slateblue r g b / 15%);
+  outline: 2px solid var(--highlight-background);
+  outline-offset: -1px;
+  background: var(--highlight-overlay)
 }
 ```
 
@@ -1208,8 +1168,8 @@ function Example() {
 
 ```css
 .react-aria-ListBoxItem[data-drop-target] {
-  box-shadow: inset 0 0 0 2px var(--highlight-background);
-  background: rgb(from slateblue r g b / 15%);
+  outline: 2px solid var(--highlight-background);
+  background: var(--highlight-overlay)
 }
 ```
