+++ packages/react-aria-components/docs/Modal.mdx
@@ -78,6 +78,8 @@ import {DialogTrigger, Modal, Dialog, Button, Heading, TextField, Label, Input} 
 ```
 
 ```css
+@import "@react-aria/example-theme/index.css";
+
 .react-aria-ModalOverlay {
   position: fixed;
   top: 0;
@@ -101,8 +103,8 @@ import {DialogTrigger, Modal, Dialog, Button, Heading, TextField, Label, Input} 
 .react-aria-Modal {
   box-shadow: 0 8px 20px rgba(0 0 0 / 0.1);
   border-radius: 6px;
-  background: var(--page-background);
-  border: 1px solid var(--spectrum-global-color-gray-300);
+  background: var(--overlay-background);
+  border: 1px solid var(--gray-400);
   outline: none;
   max-width: 300px;
 
@@ -134,7 +136,6 @@ import {DialogTrigger, Modal, Dialog, Button, Heading, TextField, Label, Input} 
     transform: scale(1);
   }
 }
-
 ```
 
 </details>
@@ -257,9 +258,9 @@ import {ModalOverlay} from 'react-aria-components';
   bottom: 0;
   right: 0;
   width: 300px;
-  background: var(--page-background);
+  background: var(--overlay-background);
   outline: none;
-  border-left: 1px solid var(--spectrum-global-color-gray-400);
+  border-left: 1px solid var(--border-color);
   box-shadow: -8px 0 20px rgba(0 0 0 / 0.1);
 
   &[data-entering] {
