+++ packages/react-aria-components/docs/RadioGroup.mdx
@@ -64,16 +64,9 @@ import {RadioGroup, Radio, Label} from 'react-aria-components';
 ```
 
 ```css
-.react-aria-RadioGroup {
-  --label-color: var(--spectrum-alias-text-color);
-  --deselected-color: gray;
-  --deselected-color-pressed: dimgray;
-  --background-color: var(--spectrum-global-color-gray-50);
-  --selected-color: slateblue;
-  --selected-color-pressed: lch(from slateblue calc(l - 10%) c h);
-  --invalid-color: var(--spectrum-global-color-red-600);
-  --invalid-color-pressed: var(--spectrum-global-color-red-700);
+@import "@react-aria/example-theme/index.css";
 
+.react-aria-RadioGroup {
   display: flex;
   flex-direction: column;
   gap: 8px;
@@ -84,7 +77,8 @@ import {RadioGroup, Radio, Label} from 'react-aria-components';
   align-items: center;
   gap: 0.571rem;
   font-size: 1.143rem;
-  color: var(--label-color);
+  color: var(--text-color);
+  forced-color-adjust: none;
 
   &:before {
     content: '';
@@ -92,45 +86,30 @@ import {RadioGroup, Radio, Label} from 'react-aria-components';
     width: 1.286rem;
     height: 1.286rem;
     box-sizing: border-box;
-    border: 0.143rem solid var(--deselected-color);
-    background: var(--background-color);
+    border: 0.143rem solid var(--border-color);
+    background: var(--field-background);
     border-radius: 1.286rem;
     transition: all 200ms;
   }
 
   &[data-pressed]:before {
-    border-color: var(--deselected-color-pressed);
+    border-color: var(--border-color-pressed);
   }
 
   &[data-selected] {
     &:before {
-      border-color: var(--selected-color);
+      border-color: var(--highlight-background);
       border-width: 0.429rem;
     }
 
     &[data-pressed]:before {
-      border-color: var(--selected-color-pressed);
+      border-color: var(--highlight-background-pressed);
     }
   }
 
   &[data-focus-visible]:before {
-    box-shadow: 0 0 0 2px var(--spectrum-alias-background-color-default), 0 0 0 4px var(--selected-color);
-  }
-}
-
-@media (forced-colors: active) {
-  .react-aria-RadioGroup {
-    forced-color-adjust: none;
-
-    --label-color: ButtonText;
-    --deselected-color: ButtonBorder;
-    --deselected-color-pressed: ButtonBorder;
-    --selected-color: Highlight;
-    --selected-color-pressed: Highlight;
-    --background-color: HighlightText;
-    --invalid-color: LinkText;
-    --invalid-color-pressed: LinkText;
-    --spectrum-alias-background-color-default: Canvas;
+    outline: 2px solid var(--focus-ring-color);
+    outline-offset: 2px;
   }
 }
 ```
@@ -421,17 +400,10 @@ The entire RadioGroup can be disabled with the `isDisabled` prop.
 ```css
 .react-aria-Radio {
   &[data-disabled] {
-    opacity: 0.4;
-  }
-}
+    color: var(--text-color-disabled);
 
-@media (forced-colors: active) {
-  .react-aria-Radio {
-    &[data-disabled] {
-      opacity: 1;
-      --deselected-color: GrayText;
-      --selected-color: GrayText;
-      --label-color: GrayText;
+    &:before {
+      border-color: var(--border-color-disabled);
     }
   }
 }
