+++ packages/react-aria-components/docs/RangeCalendar.mdx
@@ -62,20 +62,9 @@ import {RangeCalendar, Heading, Button, CalendarGrid, CalendarCell} from 'react-
 ```
 
 ```css
-.react-aria-RangeCalendar {
-  --highlight-background: slateblue;
-  --highlight-foreground: white;
-  --button-background: var(--spectrum-global-color-gray-50);
-  --button-backgound-pressed: var(--spectrum-global-color-gray-100);
-  --button-border: var(--spectrum-global-color-gray-400);
-  --button-border-disabled: var(--spectrum-alias-border-color-disabled);
-  --text-color: var(--spectrum-alias-text-color);
-  --text-color-disabled: var(--spectrum-alias-text-color-disabled);
-  --text-color-invalid: var(--spectrum-global-color-red-600);
-  --pressed-color: var(--spectrum-global-color-gray-200);
-  --unavailable-color: var(--spectrum-global-color-red-600);
-  --invalid-color: var(--spectrum-global-color-static-red-600);
+@import "@react-aria/example-theme/index.css";
 
+.react-aria-RangeCalendar {
   width: fit-content;
   max-width: 100%;
   color: var(--text-color);
@@ -114,17 +103,19 @@ import {RangeCalendar, Heading, Button, CalendarGrid, CalendarCell} from 'react-
     border-radius: 6px;
     cursor: default;
     outline: none;
+    forced-color-adjust: none;
 
     &[data-outside-month] {
       display: none;
     }
 
     &[data-pressed] {
-      background: var(--pressed-color);
+      background: var(--gray-100);
     }
 
     &[data-focus-visible] {
-      box-shadow: inset 0 0 0 2px var(--highlight-background);
+      outline: 2px solid var(--highlight-background);
+      outline-offset: -2px;
     }
 
     &[data-selected] {
@@ -133,7 +124,8 @@ import {RangeCalendar, Heading, Button, CalendarGrid, CalendarCell} from 'react-
       border-radius: 0;
 
       &[data-focus-visible] {
-        box-shadow: inset 0 0 0 1px var(--highlight-background), inset 0 0 0 3px var(--highlight-foreground);
+        outline-color: var(--highlight-foreground);
+        outline-offset: -3px;
       }
     }
 
@@ -146,22 +138,6 @@ import {RangeCalendar, Heading, Button, CalendarGrid, CalendarCell} from 'react-
       border-start-end-radius: 6px;
       border-end-end-radius: 6px;
     }
-  }
-}
-
-@media (forced-colors: active) {
-  .react-aria-RangeCalendar {
-    forced-color-adjust: none;
-
-    --highlight-background: Highlight;
-    --highlight-foreground: HighlightText;
-    --text-color: ButtonText;
-    --text-color-disabled: GrayText;
-    --text-color-invalid: LinkText;
-    --unavailable-color: GrayText;
-    --invalid-color: LinkText;
-    --pressed-color: Canvas;
-    --spectrum-alias-background-color-default: Canvas;
   }
 }
 ```
@@ -471,7 +447,7 @@ function Example() {
 
     &[data-unavailable] {
       text-decoration: line-through;
-      color: var(--unavailable-color);
+      color: var(--invalid-color);
     }
 
     &[data-invalid] {
@@ -538,7 +514,7 @@ function Example() {
 .react-aria-RangeCalendar {
   [slot=errorMessage] {
     font-size: 12px;
-    color: var(--text-color-invalid);
+    color: var(--invalid-color);
   }
 }
 ```
