+++ packages/react-aria-components/docs/Switch.mdx
@@ -54,27 +54,21 @@ import {Switch} from 'react-aria-components';
   <summary style={{fontWeight: 'bold'}}><ChevronRight size="S" /> Show CSS</summary>
 
 ```css
-.react-aria-Switch {
-  --label-color: var(--spectrum-alias-text-color);
-  --deselected-color: gray;
-  --deselected-color-pressed: dimgray;
-  --selected-color: slateblue;
-  --selected-color-pressed: lch(from slateblue calc(l - 10%) c h);
-  --track-color: var(--spectrum-alias-background-color-default);
-  --indicator-color: var(--selected-color);
-  --indicator-color-pressed: var(--selected-color-pressed);
+@import "@react-aria/example-theme/index.css";
 
+.react-aria-Switch {
   display: flex;
   align-items: center;
   gap: 0.571rem;
   font-size: 1.143rem;
-  color: var(--label-color);
+  color: var(--text-color);
+  forced-color-adjust: none;
 
   .indicator {
     width: 2rem;
     height: 1.143rem;
-    border: 2px solid var(--deselected-color);
-    background: var(--track-color);
+    border: 2px solid var(--border-color);
+    background: var(--background-color);
     border-radius: 1.143rem;
     transition: all 200ms;
 
@@ -84,57 +78,42 @@ import {Switch} from 'react-aria-components';
       margin: 0.143rem;
       width: 0.857rem;
       height: 0.857rem;
-      background: var(--indicator-color);
+      background: var(--highlight-background);
       border-radius: 16px;
       transition: all 200ms;
     }
   }
 
   &[data-pressed] .indicator {
-    border-color: var(--deselected-color-pressed);
+    border-color: var(--border-color-pressed);
 
     &:before {
-      background: var(--indicator-color-pressed);
+      background: var(--highlight-background-pressed);
     }
   }
 
   &[data-selected] {
     .indicator {
-      border-color: var(--selected-color);
-      background: var(--selected-color);
+      border-color: var(--highlight-background);
+      background: var(--highlight-background);
 
       &:before {
-        background: var(--track-color);
+        background: var(--field-background);
         transform: translateX(100%);
       }
     }
 
     &[data-pressed] {
       .indicator {
-        border-color: var(--selected-color-pressed);
-        background: var(--selected-color-pressed);
+        border-color: var(--highlight-background-pressed);
+        background: var(--highlight-background-pressed);
       }
     }
   }
 
   &[data-focus-visible] .indicator {
-    box-shadow: 0 0 0 2px var(--spectrum-alias-background-color-default), 0 0 0 4px var(--selected-color);
-  }
-}
-
-@media (forced-colors: active) {
-  .react-aria-Switch {
-    forced-color-adjust: none;
-
-    --label-color: ButtonText;
-    --deselected-color: ButtonBorder;
-    --deselected-color-pressed: ButtonBorder;
-    --indicator-color: ButtonText;
-    --indicator-color-pressed: ButtonText;
-    --track-color: ButtonFace;
-    --selected-color: Highlight;
-    --selected-color-pressed: Highlight;
-    --spectrum-alias-background-color-default: Canvas;
+    outline: 2px solid var(--focus-ring-color);
+    outline-offset: 2px;
   }
 }
 ```
@@ -238,18 +217,14 @@ Switches can be disabled using the `isDisabled` prop.
 ```css
 .react-aria-Switch {
   &[data-disabled] {
-    opacity: 0.4;
-  }
-}
+    color: var(--text-color-disabled);
 
-@media (forced-colors: active) {
-  .react-aria-Switch {
-    &[data-disabled] {
-      opacity: 1;
-      --deselected-color: GrayText;
-      --selected-color: GrayText;
-      --label-color: GrayText;
-      --indicator-color: GrayText;
+    .indicator {
+      border-color: var(--border-color-disabled);
+
+      &:before {
+        background: var(--border-color-disabled);
+      }
     }
   }
 }
