+++ packages/react-aria-components/docs/TextField.mdx
@@ -58,15 +58,9 @@ import {TextField, Label, Input} from 'react-aria-components';
 ```
 
 ```css
-.react-aria-TextField {
-  --field-border: var(--spectrum-alias-border-color);
-  --field-border-disabled: var(--spectrum-alias-border-color-disabled);
-  --field-background: var(--spectrum-global-color-gray-50);
-  --text-color: var(--spectrum-alias-text-color);
-  --text-color-disabled: var(--spectrum-alias-text-color-disabled);
-  --focus-ring-color: slateblue;
-  --invalid-color: var(--spectrum-global-color-red-600);
+@import "@react-aria/example-theme/index.css";
 
+.react-aria-TextField {
   display: flex;
   flex-direction: column;
   width: fit-content;
@@ -75,29 +69,16 @@ import {TextField, Label, Input} from 'react-aria-components';
   .react-aria-TextArea {
     padding: 0.286rem;
     margin: 0;
-    border: 1px solid var(--field-border);
+    border: 1px solid var(--border-color);
     border-radius: 6px;
     background: var(--field-background);
     font-size: 1.143rem;
-    color: var(--text-color);
+    color: var(--field-text-color);
 
     &[data-focused] {
-      outline: none;
-      border-color: var(--focus-ring-color);
-      box-shadow: 0 0 0 1px var(--focus-ring-color);
+      outline: 2px solid var(--focus-ring-color);
+      outline-offset: -1px;
     }
-  }
-}
-
-@media (forced-colors: active) {
-  .react-aria-TextField {
-    --field-border: ButtonBorder;
-    --field-border-disabled: GrayText;
-    --field-background: Field;
-    --text-color: FieldText;
-    --text-color-disabled: GrayText;
-    --focus-ring-color: Highlight;
-    --invalid-color: LinkText;
   }
 }
 ```
@@ -278,12 +259,6 @@ import {Form, FieldError, Button} from 'react-aria-components';
   .react-aria-TextArea {
     &[data-invalid] {
       border-color: var(--invalid-color);
-
-      &[data-focused] {
-        outline: none;
-        border-color: var(--focus-ring-color);
-        box-shadow: 0 0 0 1px var(--focus-ring-color);
-      }
     }
   }
 
@@ -340,7 +315,7 @@ A TextField can be disabled using the `isDisabled` prop.
   .react-aria-Input,
   .react-aria-TextArea {
     &[data-disabled] {
-      border-color: var(--field-border-disabled);
+      border-color: var(--border-color-disabled);
       color: var(--text-color-disabled);
     }
   }
@@ -530,15 +505,9 @@ fieldset {
 }
 
 .invalid {
-  color: var(--spectrum-global-color-red-600);
+  color: var(--invalid-color);
   margin-top: 1em;
   display: block;
-}
-
-@media (forced-colors: active) {
-  .invalid {
-    color: LinkText;
-  }
 }
 ```
