+++ packages/react-aria-components/docs/TimeField.mdx
@@ -60,30 +60,22 @@ import {TimeField, Label, DateInput, DateSegment} from 'react-aria-components';
 ```
 
 ```css
-.react-aria-TimeField {
-  --field-border: var(--spectrum-global-color-gray-400);
-  --field-background: var(--spectrum-global-color-gray-50);
-  --text-color: var(--spectrum-alias-text-color);
-  --text-color-placeholder: var(--spectrum-global-color-gray-700);
-  --text-color-invalid: var(--spectrum-global-color-red-600);
-  --highlight-background: slateblue;
-  --highlight-foreground: white;
-  --highlight-background-invalid: var(--spectrum-global-color-static-red-600);
-}
+@import "@react-aria/example-theme/index.css";
 
 .react-aria-DateInput {
   display: flex;
   padding: 4px;
-  border: 1px solid var(--field-border);
+  border: 1px solid var(--border-color);
   border-radius: 6px;
   background: var(--field-background);
   width: fit-content;
   min-width: 150px;
   white-space: nowrap;
+  forced-color-adjust: none;
 
   &[data-focus-within] {
-    border-color: var(--highlight-background);
-    box-shadow: 0 0 0 1px var(--highlight-background);
+    outline: 2px solid var(--focus-ring-color);
+    outline-offset: -1px;
   }
 }
 
@@ -108,21 +100,6 @@ import {TimeField, Label, DateInput, DateSegment} from 'react-aria-components';
     outline: none;
     border-radius: 4px;
     caret-color: transparent;
-  }
-}
-
-@media (forced-colors: active) {
-  .react-aria-TimeField {
-    forced-color-adjust: none;
-    --text-color-invalid: LinkText;
-    --field-border: ButtonBorder;
-    --field-background: Field;
-    --text-color: FieldText;
-    --text-color-placeholder: FieldText;
-    --text-color-invalid: LinkText;
-    --highlight-background: Highlight;
-    --highlight-foreground: HighlightText;
-    --highlight-background-invalid: LinkText;
   }
 }
 ```
@@ -368,7 +345,7 @@ import {Form, FieldError, Button} from 'react-aria-components';
 ```css
 .react-aria-DateSegment {
   &[data-invalid] {
-    color: var(--text-color-invalid);
+    color: var(--invalid-color);
 
     &:focus {
       background: var(--highlight-background-invalid);
@@ -379,7 +356,7 @@ import {Form, FieldError, Button} from 'react-aria-components';
 
 .react-aria-FieldError {
   font-size: 12px;
-  color: var(--text-color-invalid);
+  color: var(--invalid-color);
 }
 ```
 
@@ -657,6 +634,24 @@ function FieldGroup({title, children, errorMessage}: FieldGroupProps) {
   <MyTimeField label="End" defaultValue={new Time(9)} />
 </FieldGroup>
 ```
+
+<details>
+  <summary style={{fontWeight: 'bold'}}><ChevronRight size="S" /> Show CSS</summary>
+
+```css
+fieldset {
+  padding: 1.5em;
+  width: fit-content;
+}
+
+.invalid {
+  color: var(--invalid-color);
+  margin-top: 1em;
+  display: block;
+}
+```
+
+</details>
 
 ### Custom children
