+++ packages/react-aria-components/docs/react-aria-components.mdx
@@ -111,18 +111,15 @@ import {Select, SelectValue, Label, Button, Popover, ListBox, ListBoxItem} from 
   <summary style={{fontWeight: 'bold'}}><ChevronRight size="S" /> Show CSS</summary>
 
 ```css
-.react-aria-Select {
-  --border-color: var(--spectrum-alias-border-color);
-  --border-color-disabled: var(--spectrum-alias-border-color-disabled);
-  --text-color: var(--spectrum-alias-text-color);
-  --text-color-disabled: var(--spectrum-alias-text-color-disabled);
-  --focus-ring-color: slateblue;
+@import "@react-aria/example-theme/index.css";
 
+.react-aria-Select {
   .react-aria-Button {
     color: var(--text-color);
-    background: var(--spectrum-global-color-gray-50);
+    background: var(--button-background);
     border: 1px solid var(--border-color);
     box-shadow: 0 1px 2px rgba(0 0 0 / 0.1);
+    forced-color-adjust: none;
     border-radius: 6px;
     appearance: none;
     vertical-align: middle;
@@ -135,82 +132,40 @@ import {Select, SelectValue, Label, Button, Popover, ListBox, ListBoxItem} from 
     max-width: 250px;
 
     &[data-focus-visible] {
-      border-color: var(--focus-ring-color);
-      box-shadow: 0 0 0 1px var(--focus-ring-color);
+      outline: 2px solid var(--focus-ring-color);
+      outline-offset: -1px;
     }
 
     &[data-pressed] {
-      background: var(--spectrum-global-color-gray-150);
-    }
-
-    &[data-disabled] {
-      border-color: var(--border-color-disabled);
-      color: var(--text-color-disabled);
-      & span[aria-hidden] {
-        background: var(--border-color-disabled);
-      }
-
-      .react-aria-SelectValue {
-        &[data-placeholder] {
-          color: var(--text-color-disabled);
-        }
-      }
+      background: var(--button-background-pressed);
     }
   }
 
   .react-aria-SelectValue {
     &[data-placeholder] {
       font-style: italic;
-      color: var(--spectrum-global-color-gray-700);
-    }
-
-    & [slot=description] {
-      display: none;
+      color: var(--text-color-placeholder);
     }
   }
 
-  & span[aria-hidden] {
+  span[aria-hidden] {
     width: 1.5rem;
     line-height: 1.375rem;
     margin-left: 1rem;
     padding: 1px;
-    background: slateblue;
-    color: white;
+    background: var(--highlight-background);
+    color: var(--highlight-foreground);
     border-radius: 4px;
     font-size: 0.857rem;
-  }
-
-  [slot=description] {
-    font-size: 12px;
-  }
-
-  [slot=errorMessage] {
-    font-size: 12px;
-    color: var(--spectrum-global-color-red-600);
   }
 }
 
 .react-aria-ListBox {
-  --highlight-background: slateblue;
-  --highlight-foreground: white;
-  --text-color: var(--spectrum-alias-text-color);
-  --text-color-disabled: var(--spectrum-alias-text-color-disabled);
-
   max-height: inherit;
   box-sizing: border-box;
   overflow: auto;
   padding: 2px;
   outline: none;
-
-  .react-aria-Section:not(:first-child) {
-    margin-top: 12px;
-  }
-
-  .react-aria-Header {
-    font-size: 1.143rem;
-    font-weight: bold;
-    padding: 0 0.571rem 0 1.571rem;
-  }
 
   .react-aria-ListBoxItem {
     margin: 2px;
@@ -223,6 +178,7 @@ import {Select, SelectValue, Label, Button, Popover, ListBox, ListBoxItem} from 
     position: relative;
     display: flex;
     flex-direction: column;
+    forced-color-adjust: none;
 
     &[data-selected] {
       font-weight: 600;
@@ -242,33 +198,19 @@ import {Select, SelectValue, Label, Button, Popover, ListBox, ListBoxItem} from 
       background: var(--highlight-background);
       color: var(--highlight-foreground);
     }
-
-    &[data-disabled] {
-      color: var(--text-color-disabled);
-    }
-
-    [slot=label] {
-      font-weight: bold;
-    }
-
-    [slot=description] {
-      font-size: small;
-    }
   }
 }
 
 .react-aria-Popover {
-  --background-color: var(--page-background);
-  --border-color: var(--spectrum-global-color-gray-400);
+  --background-color: var(--overlay-background);
 
-  border: 1px solid var(--border-color);
-  min-width: var(--trigger-width);
-  max-width: 250px;
-  box-sizing: border-box;
+  border: 1px solid var(--gray-200);
   box-shadow: 0 8px 20px rgba(0 0 0 / 0.1);
   border-radius: 6px;
   background: var(--background-color);
   outline: none;
+  max-width: 250px;
+  box-sizing: border-box;
 
   &[data-placement=top] {
     --origin: translateY(8px);
@@ -296,31 +238,6 @@ import {Select, SelectValue, Label, Button, Popover, ListBox, ListBoxItem} from 
   to {
     transform: translateY(0);
     opacity: 1;
-  }
-}
-
-@media (forced-colors: active) {
-  .react-aria-Select {
-    --border-color: ButtonBorder;
-    --border-color-disabled: GrayText;
-    --text-color: ButtonText;
-    --text-color-disabled: GrayText;
-    --focus-ring-color: Highlight;
-
-    .react-aria-Button[data-disabled] span[aria-hidden] {
-      background: transparent;
-    }
-  }
-
-  .react-aria-ListBox {
-    forced-color-adjust: none;
-
-    --highlight-background: Highlight;
-    --highlight-foreground: HighlightText;
-    --border-color: ButtonBorder;
-    --background-color: ButtonFace;
-    --text-color: ButtonText;
-    --text-color-disabled: GrayText;
   }
 }
 ```
